submodules/DiffusionEngine/requirements.txt
diffusers==0.32.2
transformers==4.49.0
hydra-core=1.3.2
omegaconf==2.3.0